Have done very much work with authentication
Firebase is now connected to the project, newUser works sometimes, it is very weird. Using custom hooks for creating user to the database, we also check that this user do not exist with same email or username. Another hook for printing out error messages that might occur. I also seperated the login/signup so we render one component at a time instead of using if-statement in one component.