Created profile page and some more
Have created a profile page which should display information about the person and also the images.
For now I have added the header for the user which should display information about the user, like: The name, follower, following and posts. Also a name they can change that is shown in the profile with a description.