user can now log out
You can now log out and sent back to homepage. The user will be redirected to the auth page if they aren't authenticated and not use the search bar to glitch through. using the file authStore to authenticate if a user is loged in or not by looking at a local file.