Some small changes
Added a title when the user signed in so we could see which user is signed in. Will have to work on that a bit more to make it better looking, but it's good enough for now. Also changed the signIn method from username to email since firebase uses email to sign in and not username.