Added edit profile
Now a pop-up exist to fill in information about the user, but everything is not working yet. The information isn't saved but it do get the information from the database like the name and so on. Also edited all users to have a new information which is just the bold text above the biography