Update user information on database
Have added custom hook to update the user information when we click save. Update the profilepic if there is one and upload it to the database and then fetch it so we can set it locally. Also update all the user infos and if they haven't put any then use the ones already existing locally because we don't want to delete if they forgot to edit back something.
PROBLEM: We cannot run this code since we cannot update the image in the database and we make a call that fails. So I will start working with that and fetching the information about the user.