Fetching profile information works
I have now added a global hook (with zustand) for the users profile information, so we can now get the information about the user at all places. To set these values I have added a custom hook to fetch that information and it checks that the user exists, that happens in the query. The document fetched, if we can find it, is then set in the global hook, so we can use that information in other places. So we can get the name, bio and so on.