Fetching name by the URL
We now can fetch the profile by the name where we get the name from the URL. This is because we use "username" as a dynamic URL and can then fetch it. We do have to make it to lower case and do a search with that. Also added a new variable to the user which is the username in lower case. This helps with the searching of the user since if the username is with a big letter it can cause problem with the query.