Now the update works for the user
The users info can now be updated in the profile. The information is also fetched and updated directly. If the users changes their bio it will update it automatic and display it. But if the user changes their username it will change it and also have a useEffect function that checks this. When the username changes compared to the url it will automaticly reload hte page to the new user profile.
Also did some more small changes