We can now add and remove users from firebase
The following array on firebase can now be updated and shows people we follow. But we cannot see followers yet. When this is merged with main we will get the profile information from the different places with hook that was implemented there so we can update both users at the same time