Some fixes and buggs
The user can now sign in and out without the application crashing on them. This was because the user was signed out before the application hade changed component which caused some component to crash because they were trying to display "undefined"-objects. Also changed so that we navigate to our profile with "usernameLower" instead of "username" because this could've caused confusion to searching with small and big letters