Following and unfollowing users works
We can now follow and unfollow users. The action is also setting the array on the could aswell locally. So when we hit follow we will update the document online to follow that user but also making sure that we update that user to have us as a follower. We update locally also with the local files but the global hook too so we will see the changes in real time.