We can now fetch users by searching
I have implmented a modal that pops up with a search bar that lets the user write a name, although we can only get results if we write in the correct username. (Maybe work on it more but for now it is enough for me). To get the result when we write in a search on the modal we use the custom hook which searches for the user. Next step is to display the user result