Users can now follow and unfollow through the search pop-up
We use the same component for suggested users to display search results, the users. We can also search for letters and get everyone that starts with that letter for better search functionallity. We have also changed a bit of code in the custom hook for follow and unfollow becuase if we are on our own profile and follow someone we want our profile to be updated with the correct information.