Uploading pictures
We can now upload pictures to the database, but there is a limit of the images and for some reasons it is very small. We also have another global hook to store the posts we create, we will have to fetch them later if we want to be able to get all the posts from the database. We also have a cretePost component which opens a window to upload an image and write a text.
This one will make several calls to the documents and make documents for each image with its corresponding information. The image is also uploaded to the cloud storage.