Deleting posts now works
Users can delete posts and if they do it also updates the profile! It removes the post from the storage on the database and also the document which saves the image with its corresponding info. The reference is also deleted from the users post array. We also update the profile when this happens and updates the number of posts aswell as the posts shown on the profile