We can now comment
Comments are now shown on the post and being updated when we post a new comment. Although we do not get the user information yet. We also have changed if we are seeing info about number of posts depending if we are on the homepage or in a post. We add the comment by using a custom hook which does the calls to firebase and so on. THat hook also calls our global variable in postsStore.js which handles if a change happens we want to update the ui.