Fixed image sizes in different screen sizes
The images can now be in standing format but also in landscape and look good. I did as good as I could so that images and comment section would be working together seemingly in different sizes of screens, which is both computers and phones.
Also added a if-statement in the post comment hook to check that the comment isn't empty