Added likes hook
This hook should let the user like and unlike images. Fetching the number of like in the database, which is stored in each post we can set the number of likes. If the user has liked this post we set the value and then we can unlike and like images while also updating the firebase.