Changed suggestedUser Component
Now the suggested user will set the global state of the user when we press their profile to navigate there. Before it navigated without setting the state, which caused an error since we couldn't fetch a user because we didn't have one set.
But now we can navigate to users profile by clicking them and not getting error